Change debug app to sample app, demonstrating deep links, app links and starting RoverActivity
