--- debug-app/src/main/java/io/rover/app/debug/DebugApplication.kt
@@ -1,25 +1,0 @@
-package io.rover.app.debug
-
-import android.app.Application
-import com.appspector.sdk.AppSpector
-import io.rover.sdk.Rover
-import timber.log.Timber
-
-class DebugApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-
-        if (BuildConfig.DEBUG) {
-            Timber.plant(Timber.DebugTree())
-        }
-
-        // You can start all monitors
-        AppSpector
-            .build(this)
-            .withDefaultMonitors()
-            .run("android_NWJhNTMwYTMtM2RlOS00MzFlLWEzMWMtMTU5YWUzNjBlZjdk");
-
-        Rover.installSaneGlobalHttpCache(this)
-        Rover.initialize(this, getString(R.string.rover_api_token))
-    }
-}
