--- debug-app/src/main/res/values/strings.xml
@@ -1,7 +1,0 @@
-<resources>
-    <string name="app_name">Rover Debug</string>
-    <string name="title_home">Home</string>
-    <string name="title_notifications">Notification Center</string>
-    <string name="title_geo">Geo</string>
-    <string name="title_settings">Settings</string>
-</resources>
