--- debug-app/src/main/res/xml/network_security_config.xml
@@ -1,8 +1,0 @@
-<network-security-config>
-    <debug-overrides>
-        <trust-anchors>
-            <!-- Trust user added CAs while debuggable only -->
-            <certificates src="user" />
-        </trust-anchors>
-    </debug-overrides>
-</network-security-config>
