+++ sample-app/proguard-rules.pro
