+++ sample-app/src/main/AndroidManifest.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="io.rover.app.sample">
+
+    <application
+        android:name="io.rover.app.sample.SampleApplication"
+        android:allowBackup="true"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:supportsRtl="true"
+        android:theme="@style/AppTheme">
+
+        <!-- Registering RoverActivity in order to display experiences -->
+        <activity
+            android:name="io.rover.sdk.ui.containers.RoverActivity"
+            android:theme="@style/AppTheme.NoActionBar" />
+
+        <activity
+            android:name="io.rover.app.sample.SampleMainActivity"
+            android:label="@string/app_name">
+
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+
+            <!-- for deep links to open experiences -->
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT"/>
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <data android:scheme="@string/uri_scheme" />
+            </intent-filter>
+
+            <!-- for http universal links/app links to open experiences  -->
+            <intent-filter android:autoVerify="true">
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT"/>
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <data android:scheme="http" android:host="@string/associated_domain" />
+            </intent-filter>
+
+            <!-- for https universal links/app links to open experiences -->
+            <intent-filter android:autoVerify="true">
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT"/>
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <data android:scheme="https" android:host="@string/associated_domain" />
+            </intent-filter>
+
+        </activity>
+    </application>
+
+</manifest>
