+++ sample-app/src/main/java/io/rover/app/sample/SampleApplication.kt
@@ -1,0 +1,15 @@
+package io.rover.app.sample
+
+import android.app.Application
+import io.rover.sdk.Rover
+
+const val ROVER_API_TOKEN = "BLANK"
+
+class SampleApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        Rover.installSaneGlobalHttpCache(this)
+        Rover.initialize(this, ROVER_API_TOKEN)
+    }
+}
