+++ sample-app/src/main/res/layout/activity_sample_main.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.rover.app.debug.DebugMainActivity">
+    tools:context="io.rover.app.sample.SampleMainActivity">
 </android.support.constraint.ConstraintLayout>
