+++ sample-app/src/main/res/values/constants.xml
@@ -1,0 +1,5 @@
+<resources>
+    <string name="uri_scheme">blank</string>
+    <string name="associated_domain">blank</string>
+    <string name="experience_id">blank</string>
+</resources>
