+++ settings.gradle
@@ -1,1 +1,1 @@
-include 'sdk', 'debug-app'
+include 'sdk', 'sample-app'
